If we already have a bearer token (such as in tests), don't delete the bearer token

+++ main.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"io/fs"
 	"log/slog"
 	"math"
 	"os"
@@ -286,14 +287,20 @@ func run(ctx context.Context, k *koanf.Koanf) error {
 		h = h.withKerberosKeytabAuth(k.String("vault.kerberosKeytabPath"), k.String("vault.kerberosPrincipal"))
 	}
 
+	deleteBearerTokenAfterRun := false // We only want to delete a bearer token if we created it
+	if _, err := os.Stat(h.outFile); err != nil && errors.Is(err, fs.ErrNotExist) {
+		deleteBearerTokenAfterRun = true
+	}
 	tok, err := h.getToken(ctx, k.String("vault.experiment"), k.String("vault.role"))
 	if err != nil {
 		return fmt.Errorf("error getting and validating token: %w", err)
 	}
-	defer func() {
-		os.Remove(h.outFile) // Clean up the token file after we're done
-		funcLogger.Debug("Removed bearer token file", "file", h.outFile)
-	}()
+	if deleteBearerTokenAfterRun {
+		defer func() {
+			os.Remove(h.outFile) // Clean up the token file after we're done
+			funcLogger.Debug("Removed bearer token file", "file", h.outFile)
+		}()
+	}
 
 	promDuration.WithLabelValues("getBearerToken").Set(time.Since(startGetBearerToken).Seconds())
 	promDuration.WithLabelValues("setup").Set(time.Since(startSetup).Seconds())
